refactor: point matrix-rain consumers at the file (no barrel)
Repo doesn't use index barrels; import effects/matrix-rain/matrix-rain
directly in the four consumers (2 molecules + /404 + /offline).
