+++ src/app/not-found.tsx
@@ -1,4 +1,4 @@
-import { MatrixRain } from "@/components/design-system/atoms/effects/matrix-rain";
+import { MatrixRain } from "@/components/design-system/atoms/effects/matrix-rain/matrix-rain";
 import { MatrixTerminal } from "@/components/design-system/molecules/effects/matrix-terminal";
 import { BluePillLink, RedPillLink } from "@/components/design-system/molecules/links/pills-links";
 import { tracking } from "@/types/configuration/tracking";
