+++ src/app/offline/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { MatrixRain } from "@/components/design-system/atoms/effects/matrix-rain";
+import { MatrixRain } from "@/components/design-system/atoms/effects/matrix-rain/matrix-rain";
 import { MatrixTerminal } from "@/components/design-system/molecules/effects/matrix-terminal";
 import { BluePillLink } from "@/components/design-system/molecules/links/pills-links";
 import { RedPillButton } from "@/components/design-system/molecules/buttons/pills-buttons";
